move runKibanaServer specific cli arg logic back into that file
